+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.2.3) (YYYY-MM-DD)
 Updated analytics SDK dependency.
 Updated AdColony network library.
+Updated InMobi network library.
 Updated target Android version.
 
 ## [1.2.2](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.2.2) (2016-09-15)