+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [1.2.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.2.3) (YYYY-MM-DD)
+Fixed Amazon ad events.  
 Updated analytics SDK dependency.  
 Updated AdColony network library.  
 Updated InMobi network library.