+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Fixed Amazon ad events.
 Updated analytics SDK dependency.  
 Updated AdColony network library.  
 Updated InMobi network library.  
+Updated MoPub network library.  
 Updated target Android version.  
 
 ## [1.2.2](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.2.2) (2016-09-15)