+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.2.3) (YYYY-MM-DD)
 Fixed Amazon ad events.  
 Fixed MobFox crash.  
+Fixed Vungle misreporting ad completed watch status.  
 Updated analytics SDK dependency.  
 Updated AdColony network library.  
 Updated InMobi network library.